[Ankit Rajput] Add. Introduce Service Layer for Employee Payroll Management Application

+++ EmployeePayrollApp/src/main/java/com/bridgelabz/employeepayrollapp/dto/EmployeeRequestDTO.java
@@ -4,12 +4,12 @@ package com.bridgelabz.employeepayrollapp.dto;
 // This class is used to transfer employee details from the client to the server.
 public class EmployeeRequestDTO {
     private String name;
-    private Long salary;
+    private double salary;
 
     // Constructor
     public EmployeeRequestDTO() {}
 
-    public EmployeeRequestDTO(String name, Long salary) {
+    public EmployeeRequestDTO(String name, double salary) {
 	   this.name = name;
 	   this.salary = salary;
     }
@@ -25,12 +25,12 @@ public class EmployeeRequestDTO {
     }
 
     // Getter for salary
-    public Long getSalary() {
+    public double getSalary() {
 	   return salary;
     }
 
     // Setter for salary
-    public void setSalary(Long salary) {
+    public void setSalary(double salary) {
 	   this.salary = salary;
     }
